[2.1][feat] Create JSON from card + execute `count` function
