+++ src/components/Cohort/CreationCohort/ControlPanel/ControlPanel.tsx
@@ -2,6 +2,7 @@ import React from 'react'
 import clsx from 'clsx'
 
 import { Button, Divider, Grid, Typography } from '@material-ui/core'
+import { Skeleton } from '@material-ui/lab'
 
 import ArrowBackIcon from '@material-ui/icons/ArrowBack'
 import ArrowForwardIcon from '@material-ui/icons/ArrowForward'
@@ -9,40 +10,48 @@ import UpdateSharpIcon from '@material-ui/icons/UpdateSharp'
 
 import useStyle from './styles'
 
-const ControlPanel = () => {
+import { CohortCreationCounterType } from 'types'
+
+const ControlPanel: React.FC<
+  {
+    onExecute: () => void
+  } & CohortCreationCounterType
+> = ({ includePatient, byrequest, alive, deceased, female, male, onExecute }) => {
   const classes = useStyle()
-  const includePatient = '21200'
-  const byrequest = '21200'
-  const alive = '17500'
-  const deceased = '3700'
-  const female = '14000'
-  const male = '7200'
 
   return (
     <Grid className={classes.rightPanelContainerStyle}>
       <Grid>
         <Grid container justify="center" className={classes.requestAction}>
-          <Button className={classes.requestExecution}>Exécuter la requête</Button>
+          <Button onClick={onExecute} className={classes.requestExecution}>
+            Exécuter la requête
+          </Button>
         </Grid>
 
         <Divider />
 
-        <Button>
-          <ArrowBackIcon color="action" className={classes.iconBorder} />
+        <Button
+          className={classes.actionButton}
+          startIcon={<ArrowBackIcon color="action" className={classes.iconBorder} />}
+        >
           <Typography className={classes.boldText}>Annuler</Typography>
         </Button>
 
         <Divider />
 
-        <Button>
-          <ArrowForwardIcon color="action" className={classes.iconBorder} />
+        <Button
+          className={classes.actionButton}
+          startIcon={<ArrowForwardIcon color="action" className={classes.iconBorder} />}
+        >
           <Typography className={classes.boldText}>Rétablir</Typography>
         </Button>
 
         <Divider />
 
-        <Button>
-          <UpdateSharpIcon color="action" className={classes.iconBorder} />
+        <Button
+          className={classes.actionButton}
+          startIcon={<UpdateSharpIcon color="action" className={classes.iconBorder} />}
+        >
           <Typography className={classes.boldText}>Réinitialiser</Typography>
         </Button>
       </Grid>
@@ -51,28 +60,38 @@ const ControlPanel = () => {
         <Grid container justify="space-between">
           <Typography className={clsx(classes.boldText, classes.patientTypo)}>PATIENTS INCLUS</Typography>
           <Typography className={clsx(classes.blueText, classes.boldText, classes.patientTypo)}>
-            {includePatient}
+            {includePatient === 'loading' ? <Skeleton variant="rect" width={50} height={19} /> : includePatient ?? '-'}
           </Typography>
         </Grid>
         <Grid container justify="space-between">
           <Typography className={classes.sidesMargin}>Par requête</Typography>
-          <Typography className={clsx(classes.blueText, classes.sidesMargin)}>{byrequest}</Typography>
+          <Typography className={clsx(classes.blueText, classes.sidesMargin)}>
+            {byrequest === 'loading' ? <Skeleton variant="rect" width={50} height={19} /> : byrequest ?? '-'}
+          </Typography>
         </Grid>
         <Grid container justify="space-between">
           <Typography className={classes.sidesMargin}>Patients vivants</Typography>
-          <Typography className={clsx(classes.blueText, classes.sidesMargin)}>{alive}</Typography>
+          <Typography className={clsx(classes.blueText, classes.sidesMargin)}>
+            {alive === 'loading' ? <Skeleton variant="rect" width={50} height={19} /> : alive ?? '-'}
+          </Typography>
         </Grid>
         <Grid container justify="space-between">
           <Typography className={classes.sidesMargin}>Patients décédés</Typography>
-          <Typography className={clsx(classes.blueText, classes.sidesMargin)}>{deceased}</Typography>
+          <Typography className={clsx(classes.blueText, classes.sidesMargin)}>
+            {deceased === 'loading' ? <Skeleton variant="rect" width={50} height={19} /> : deceased ?? '-'}
+          </Typography>
         </Grid>
         <Grid container justify="space-between">
           <Typography className={clsx(classes.sidesMargin)}>Nombre de femmes</Typography>
-          <Typography className={clsx(classes.blueText, classes.sidesMargin)}>{female}</Typography>
+          <Typography className={clsx(classes.blueText, classes.sidesMargin)}>
+            {female === 'loading' ? <Skeleton variant="rect" width={50} height={19} /> : female ?? '-'}
+          </Typography>
         </Grid>
         <Grid container justify="space-between">
           <Typography className={classes.sidesMargin}>Nombre d'hommes</Typography>
-          <Typography className={clsx(classes.blueText, classes.sidesMargin)}>{male}</Typography>
+          <Typography className={clsx(classes.blueText, classes.sidesMargin)}>
+            {male === 'loading' ? <Skeleton variant="rect" width={50} height={19} /> : male ?? '-'}
+          </Typography>
         </Grid>
       </Grid>
     </Grid>
